+++ upcloud/testdata/datasource_upcloud_kubernetes_plans/basic.tf
@@ -1,0 +1,1 @@
+data "upcloud_kubernetes_plans" "basic" {}